api/requirements.txt
blinker==1.9.0
boto3==1.37.22
botocore==1.37.22
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
urllib3==2.4.0
Werkzeug==3.1.3
flasgger==0.9.7.1
openai==1.75.0
numpy==2.2.4
google-genai==1.11.0